GH2883 Move ServiceInfo Model to common-service-core
