+++ common-service-core/src/main/java/org/zowe/apiml/services/ServiceInfo.java
@@ -8,13 +8,15 @@
  * Copyright Contributors to the Zowe Project.
  */
 
-package org.zowe.apiml.product.services;
+package org.zowe.apiml.services;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.netflix.appinfo.InstanceInfo;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.zowe.apiml.auth.Authentication;
 import org.zowe.apiml.config.ApiInfo;
@@ -31,6 +33,8 @@ public class ServiceInfo {
     private Apiml apiml;
     private Map<String, Instances> instances;
 
+    @NoArgsConstructor
+    @AllArgsConstructor
     @Data
     @Builder
     @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -40,6 +44,8 @@ public class ServiceInfo {
         private List<Authentication> authentication;
     }
 
+    @NoArgsConstructor
+    @AllArgsConstructor
     @Data
     @Builder
     @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -55,6 +61,8 @@ public class ServiceInfo {
         private Map<String, String> customMetadata;
     }
 
+    @NoArgsConstructor
+    @AllArgsConstructor
     @Data
     @Builder
     @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -64,6 +72,8 @@ public class ServiceInfo {
         private String homePageUrl;
     }
 
+    @NoArgsConstructor
+    @AllArgsConstructor
     @Data
     @SuperBuilder
     @EqualsAndHashCode(callSuper = true)
