+++ gateway-service/src/main/java/org/zowe/apiml/gateway/services/ServicesInfoController.java
@@ -22,13 +22,12 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.zowe.apiml.product.services.ServiceInfo;
-import org.zowe.apiml.product.services.ServicesInfoService;
+import org.zowe.apiml.services.ServiceInfo;
 
 import java.util.List;
 
-import static org.zowe.apiml.product.services.ServicesInfoService.CURRENT_VERSION;
-import static org.zowe.apiml.product.services.ServicesInfoService.VERSION_HEADER;
+import static org.zowe.apiml.gateway.services.ServicesInfoService.CURRENT_VERSION;
+import static org.zowe.apiml.gateway.services.ServicesInfoService.VERSION_HEADER;
 
 
 @RestController
